fixed eRerouter bug for vehicles with expired parking duration

+++ src/utils/common/SUMOTime.cpp
@@ -113,7 +113,7 @@ elapsedMs2string(long long int t) {
             // round to seconds
             return time2string((t / 1000) * 1000);
         } else {
-            return toString((double)t / 1000.0) + "s";
+            return toString(t / 1000.0) + "s";
         }
     } else {
         return time2string(t) + "s";
